# Parallel model tokenizer index load